+++ lib/src/default_tags/list_tag.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_bbcode/tags/tag_parser.dart';
 import 'package:bbob_dart/bbob_dart.dart' as bbob;
 
-import '../flutter_renderer.dart';
+import '../../flutter_bbcode.dart';
+import 'abstract_tags.dart';
+
 
 /// The style of the list. This is used by the list items to generate the correct style
 enum ListType { ordered, unordered }